requirements.txt
django == 4.0.1
Pillow == 9.0.1
pip == 22.0.3
djangorestframework == 3.13.1
django-admin-list-filter-dropdown == 1.0.3
django-admin-rangefilter == 0.8.3
djangorestframework-simplejwt == 5.0.0